% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/boxplot_context.R
\name{calculer_statistiques}
\alias{calculer_statistiques}
\title{Calculer les statistiques descriptives par groupe}
\usage{
calculer_statistiques(data, x_col, y_col)
}
\arguments{
\item{data}{Le data.frame des données à plot.}

\item{x_col}{Nom de la colonne de groupement.}

\item{y_col}{Nom de la colonne numérique.}
}
\value{
Une liste contenant les statistiques par groupe.
}
\description{
Calculer les statistiques descriptives par groupe
}
\keyword{internal}
