% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_context.R
\name{construire_prompt_summary}
\alias{construire_prompt_summary}
\title{Construire le prompt pour le LLM}
\usage{
construire_prompt_summary(summary_obj)
}
\arguments{
\item{summary_obj}{Un objet de la classe summary_context.}
}
\value{
Une chaîne de caractères contenant le prompt.
}
\description{
Construire le prompt pour le LLM
}
\keyword{internal}
