% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_context.R
\name{get_variable_type}
\alias{get_variable_type}
\title{Identifier le type de la variable}
\usage{
get_variable_type(x)
}
\arguments{
\item{x}{une variable.}
}
\value{
Une liste contenant les statistiques par groupe.
}
\description{
Identifier le type de la variable
}
\keyword{internal}
