% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ollama_api.R
\name{interroger_ollama}
\alias{interroger_ollama}
\title{Interroger l'API Ollama}
\usage{
interroger_ollama(
  prompt,
  model = "mistral",
  endpoint = "http://localhost:11434"
)
}
\arguments{
\item{prompt}{Le prompt à envoyer au modèle}

\item{model}{Nom du modèle (par défaut "mistral")}

\item{endpoint}{URL de l'API Ollama (par défaut "http://localhost:11434")}
}
\value{
Une liste contenant titre et paragraphe
}
\description{
Interroger l'API Ollama
}
\keyword{internal}
