% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.boxplot_context}
\alias{plot.boxplot_context}
\title{Tracer le boxplot d'un objet boxplot_context}
\usage{
\method{plot}{boxplot_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe boxplot_context}

\item{...}{Arguments supplémentaires passés à ggplot2}
}
\value{
Un objet ggplot
}
\description{
Tracer le boxplot d'un objet boxplot_context
}
