% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.boxplot_context}
\alias{print.boxplot_context}
\title{Afficher l'interprétation d'un objet boxplot_context}
\usage{
\method{print}{boxplot_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe boxplot_context.}

\item{...}{Arguments supplémentaires (non utilisés).}
}
\description{
Affiche le titre généré par le LLM et l'interprétation associée,
ainsi que quelques métadonnées (variables, nombre de groupes).
}
\examples{
\dontrun{
data(iris)
bx <- boxplot_context(iris, "Species", "Sepal.Length")
print(bx)
}

}
\seealso{
[boxplot_context] pour créer l'objet et [plot.boxplot_context]
pour le tracé.
}
