% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary_context}
\alias{print.summary_context}
\title{Afficher le résumé d'un objet summary_context}
\usage{
\method{print}{summary_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe summary_context.}

\item{...}{Arguments supplémentaires (non utilisés).}
}
\description{
Affiche un tableau synthétique des variables (type et résumé),
le total de données manquantes et, si disponible, l'interprétation
produite par le LLM.
}
\seealso{
[summary_context] pour créer l'objet.
}
