% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.boxplot_context}
\alias{plot.boxplot_context}
\title{Tracer le boxplot d'un objet boxplot_context}
\usage{
\method{plot}{boxplot_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe boxplot_context.}

\item{...}{Arguments supplémentaires passés à ggplot2.}
}
\value{
Un objet ggplot.
}
\description{
Tracer le boxplot d'un objet boxplot_context
}
\details{
Nécessite le package `ggplot2`.
}
\examples{
\dontrun{
data(iris)
bx <- boxplot_context(iris, "Species", "Sepal.Length")
plot(bx)
}

}
\seealso{
[boxplot_context] pour construire l'objet et
[print.boxplot_context] pour l'affichage textuel.
}
