% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{plot.scatterplot_context}
\alias{plot.scatterplot_context}
\title{Méthode plot pour scatterplot_context}
\usage{
\method{plot}{scatterplot_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe scatterplot_context.}

\item{...}{Arguments supplémentaires passés à ggplot2.}
}
\value{
Un objet ggplot.
}
\description{
Méthode plot pour scatterplot_context
}
