% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.boxplot_context}
\alias{print.boxplot_context}
\title{Afficher l'interprétation d'un objet boxplot_context}
\usage{
\method{print}{boxplot_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe boxplot_context.}

\item{...}{Arguments supplémentaires (non utilisés).}
}
\description{
Afficher l'interprétation d'un objet boxplot_context
}
