% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.scatterplot_context}
\alias{print.scatterplot_context}
\title{Méthode print pour scatterplot_context}
\usage{
\method{print}{scatterplot_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe scatterplot_context.}

\item{...}{Arguments supplémentaires (non utilisés).}
}
\description{
Méthode print pour scatterplot_context
}
