% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/methods.R
\name{print.summary_context}
\alias{print.summary_context}
\title{Afficher le résumé d'un objet summary_context}
\usage{
\method{print}{summary_context}(x, ...)
}
\arguments{
\item{x}{Un objet de classe summary_context.}

\item{...}{Arguments supplémentaires (non utilisés).}
}
\description{
Afficher le résumé d'un objet summary_context
}
