% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_context.R
\name{summary_context}
\alias{summary_context}
\title{Créer un summary contextualisé avec analyse LLM}
\usage{
summary_context(data)
}
\arguments{
\item{data}{Un data.frame contenant les données.}
}
\value{
Un objet de classe summary_context.
}
\description{
Créer un summary contextualisé avec analyse LLM
}
